Restart the filter before starting the refocus operation
When there is a sensor fusion error, start a reset and restart video
processing once focus is finished. If the sensor fusion error is a
video based error and we haven't had any good video frames yet, stop
processing video but keep running the inertial filter until the focus
finishes.