Add a new target for rc3dksamplevis that builds a primitive version for iOS
This is a WIP. While connections are established and data is
transmitted, the display of said data is pretty horrible looking and
there is no way to manage the viewpoint in the app. Also, the
Bonjour connection mechanism seems to be more inconsistent than it is
on the desktop equivalent, despite using the same code.