Correct huge error caused during the recent merge with master.
stereo_mesh had the convention of using s2 as the current frame of
video, and s1 as the previous frame. Meanwhile, these parameters had
been renamed in stereo and . Correct this by flipping parameter order in
the appropriate places. To be followed by a patch which renames the
parameters to avoid further confusion.