Convert from Jacobians to Body Jacobians in state and observation
As an example for the acceleration measurement

    a = Rt * acc,

we current compute a_dW as.

    m4v4 dR_dW = to_rotation_matrix_jacobian(state.W.v);
    da_dW = transpose(dR_dW) * acc;

First some back ground

    da
      = Rt dacc +     dRt      * acc
      = Rt dacc -  Rt dR  * Rt * acc
      = Rt dacc - (Rt dR) * Rt * acc
      = Rt dacc - (Rt_dR) * (Rt * acc)

We can think of (Rt dR) as a skew symmetric matrix

      = Rt dacc - invskew3(Rt_dR) x (Rt * acc)
      = Rt dacc + skew3(Rt * acc) * invskew3(Rt_dRt)

where in the last line the derivative has moved to the right so that
we can have the covariance multipled in from the right as is our
convention.

    Rt_dR_dW = to_rotation_matrix_jacobian(state.W.v);
    da_dW = skew3(Rt * acc) * Rt_dR_dW

With final covariance projection like so:

    Rt * cov_acc + da_dW * cov_W