Use dt between measurements to scale image_velocity
Before this patch, we assumed the period between images was constant.
If we dropped a frame, or otherwise violated this assumption, the
image velocity tracker would be more likely to fail. Doesn't change
our benchmark performance much, because our assumption is not often
violated there.

Mean of 116 primary errors that are less than 50% is 8.01%
Histogram score (lower is better) is 126
Alternate histogram score (lower is better) is 102