Android: integrate cmake
This is complicated by Andoid's gradle-experimental and its lack of full
support for variants and flavors.  In particular filtering, dimensions
and mapping over variants seem to be un- or poorly supported.

There are a few warts still:

  o We needlessly run cmake when nothing has changed
  o We needlessly copy libtrack_* .so
  o We needlessly rebuild APKs
  o We include the debug .so in release builds (and vv) when it's not used
  o We are not using clang for the wrapper but we are for libtracker
  o Clang x86 crashes on homography in release mode

The good thing is that all you need to do to build is

  RSSDK=... MKLROOT=... ./gradlew build

But it's faster to just build one target

  RSSDK=... MKLROOT=... ./gradlew assembleRc64Release

at a time.