WIP checkpoint for integrating mapper with the filter
Currently projects features to a dictionary which is built on the fly
with a fixed size (30 features). Preliminary matching is based on
projected feature id. In sequences I do not see features reliably
redetected, and new features always seem to be further than I would
expect from the dictionary item.