Don't match outside of initial breadth_first search
Since the map is always connected (via adding an edge between the last
group to disappear and the first group to appear when all features are
dropped), the only time a node wouldn't be reached in a breadth first
search is when there is a stored map. This will be treated separately.