filter: Increase max_accel_delta to 20m/s^2
On wheeled robot datasets, we often see high frequency vibrations on
the order of 10m/s^2. We very occasionally see as high as 20m/s^2. We
should consider removing this check completely, or moving it to
sensor manager code, since it was originally intended to cope with bad
data buffers from iOS.