filter: create/use an aligned_undistorted_depth image when needed
This creates a depth buffer aligned to the color camera (w/ either
fisheye or polynomial distortion).

This allows us to use the raw depth data, but requires depth
intrinsics and color-to-depth extrinsics which are currently hard
coded, but can eventuall be read from Scene Perceptions
CameraParameters.txt and/or added to our config file.  This minimizes
the tan() and sqrt() call's needed to align the depth to the fisheye
by avoiding undistoring distortion for every pixel.  It also only
creates the aligned depth image when needed (~5% of frames).