Switch from SP_Calibration to a new calibration_json based on calibration_xml
Convert the internal representation of the JSON from SP_Calibration to
a new structure calibration_json that is build from the imu,
intrinsics and extrinsics used in calibration_xml.  When it is used to
represent a configureation of the filter, it is called
device_parameters.  These two, device_parameters and calibration_json,
could diverge if/when the JSON format gets extended to have more info
than is needed by the filter, like with XML format.

To make the extrinsics match, the calibration_xml has been extended
with Tc_var and Wc_var (now called extrinsics_var_wrt_imu_m).  These
are not written to the XML file, but we could extend the format in the
future.