mapper: only search for matches using the nodes features
Suppose we have the graph

1 -> 2 -> 3 --drift--> 5
          | -> 4

Where 3 is connected to 4 (with a good transform) and 5 (with some
drift). If we use the features of 5 in the reference frame of 3, we'll
end up with a transform from 1 to 5, but we will add the edge from 1
to 3. As a result, the relative position of 4 will now be completely
wrong. Since we use neighbors to build tf-idf histograms, in this case
we should search for closures on both 3 and 5, so we should ignore the
matches from 5 when looking for translation proposals for 3 (and vice
versa).