rc_tracker: make the status callback syncronous (like the others)
The data, status and message callbacks are now all syncronous with the
filter.  This moves the burden any copying and or thread creation
(std::async was creating a thread for *every* status callback on at
least Android), to the clients.

Android is currently the only platform that runs in ASYNC mode and
given that the data_callback was already syncronous and was calling
into Java and opengl, I'm going to leave the status callback also
calling into Java for now.