mapper: offset is calculated relative to from
The offset to add to the filter is the change in the position of the
"from" node of the loop closure before finding the loop closure
compared to after the edge has been added and the map has been
rebuilt.

Previously, since we used the global position of nodes, loaded maps
would provide very different offsets.