observation: Rotate initial gravity measurement to body frame
Qc goes from the frame of the camera (body) to the frame of the IMU,
so this usage needed Qc^-1. This should have caused problems, but our
benchmark dataset has only 180 degree rotations for Wc, so the inverse
had almost no effect (small numerical differences). On E6T, we put the
correct rotation in accelerometerTransform, which is applied before
creating the observation, and set Wc to 0,0,0.