filter: compose the imu and camera transforms for the one sensor case
Eventually, all sensors will compute observations relative to their
extrinsics, but now that we expect to have camera to imu transform,
satisfy that for multi-sensor configurations by computing the
transform between the first camera and imu. Change legacy calibrations
to store their origin transform on the imu rather than the camera,
since the camera is the origin (body) frame.