sensor_fusion: on numerical errors, call filter_initialize() rather than flush_and_reset()
flush_and_reset() tries to end the thread the filter is running on, so
on numerical errors the filter used to use std::async to call
flush_and_reset(), but this can be delayed arbitrarily and we can end
up queuing up multiple flush_and_reset()s which can end up with two
threads calling filter_initialize().

Instead of flush_and_reset() from a separate thread, we now call
filter_initialize() only.  This won't reset the queue or the thread we
are running on, but it should restore numerical stability.

This should be revisited.