sensor_fusion: separate the camera_callback from the data_callback
Replay needs the image and it wants it all the time. rc_DataCallback
only wants the data and it can get it for itself, like it already
does for rc_getPose() and rc_getFeatures().

This gets us one step closer to 1) calling the data_callback from the
gyro update and 2) to having the data callback use rc_getFeatures()
and rc_getPose() itself rather than forcing the data on it.