sensor_fusion: restart the filter on numerical resets
Relates to 2af99f48c7b7420b8f8ba2d90f307293ba29e74c. Note that this
approach preserves the last output pose, which means we are no longer
gravity aligned in the case that the numerical instability caused our
rotation to be wrong. We should think about preserving gravity
alignment and trying to maintain the last yaw and translation as an
alternative to preserving the last output pose.