rc_tracker: Provide access to the tracker and data in data callback
Sensor data is now returned in the data callback. Introduces a new
rc_Data struct, which has time, type, id, and a union of sensor
dependent data. We also now provide access to the tracker in the data
callback. Expected use will be calling rc_getPose to get the pose
associated with that data.