vis: don't draw when we have nothing to do
If nothing has changed in either the UI or the data, sleep for 5ms and
try again. Note, we have to stick with glfwPollEvents rather than
glfwWaitEvents because data might've changed in update_vertex_arrays
while we are waiting.

update_vertex_arrays now returns a bool to indicate if anything has
changed since the last call.