sensor_fusion_queue: Construct the queue with max_latency_us
Set a queue level maximum latency. This bounds the size of the queue
and also the maximum imposed latency. When new data arrives, we check
the oldest item in the queue compared to the newest seen timestamp (in
ok_to_dispatch) and if it is greater than this, we dispatch the old
data.