sensor_fusion_queue: add buffering
When buffering is enabled, we drop all received data that is older
than the newest data point - the buffer size. Once another start_ is
called, we dispatch the buffered data and proceed as normal.

Adds two tests to verify that we do not dispatch data if we don't
start the queue and to verify the buffered data is dispatched once the
queue is started.