capture: replace packaged_task with a std::unique_ptr to sensor_data
Now that sensor_data is an all-encompasing type, we don't need to use
packaged_task to encapsulate the write operations. We can instead
directly store the copies of the sensor data in a queue and have one
unified write function which is called when we pop things out of the
queue.