sensor_fusion_queue: Replace ELIMINATE_DROPS with MINIMIZE_DROPS
MINIMIZE_DROPS is the same as ELIMINATE_DROPS (it waits to have one of
each sensor and then dispatches), but it allows for autoconfiguration
by not waiting for the first item of each type. Once the queue has
seen a sample from a sensor, it will wait to always have one sample
from that sensor before deciding if it is ok to dispatch.