vec4: avoid conflicting #define of I from <complex.h> (only used by Eigen)
On Ubuntu <complex.h> #define's I in a way that conflicts with many of
our uses of I as local variable or function argument.  Rather than fix
this in our code and all of our dependencies (which also have
problems) we can just use <complex> which doesn't define I.  When
using LAPACKe we get <complex.h> via Eigen and we only support
including Eigen from vec4.h, so we can just #undef it there.