state: centralize all the estimate_* bools into state_branch
This makes the unit of "disabling estimation" the "the branch", and
restructures the tree of states to reflect that.  It adds a branch for
the non_orientation states, and translation states.  These can
generically be easily disabled and enabled, once we make the
reset()ing of the translation states and saving away of the bias
variances generic features.

This makes the state_branch structure mostly static, except for
features and groups, and which means that the order of states is
stable by construction rather than relying on the order of push_back's
that used to be in two places.