tracker: tune fast / outlier (m-estimator) thresholds
This attempts to tune the tracker to avoid matching corners that are
not nearly as strong as the detected corners.

We might be better of making this threshold a relative allowed
reduction in strength instead of an absolute threshold.

This make the patch matching require a stronger match.  I've see the
corner of lights match a flat (in color) cubical wall which is clearly
too flexible.

This also make is drop features sooner instead of persisting with poor
matches.