vec4: rename v_map() and m_map() to just map()
The idea is that map() returns an Eigen::Map for whatever its argument
is, just like the .map() method does on matrix, but it is restricted
to cases like arrays and soon std::vector, where we can't add methods.
This is analogous to the reasons for both std::begin() and .begin().