rc_tracker: rc_getPose returns a rc_PoseTime instead of an rc_Pose
We previously assumed that rc_getPose would be called from the data
callback and if you wanted the time you would check data->time_us.
However, for camera packets data->time_us was not the time of the pose
because the timestamp for camera packets corresponds to the start of
exposure. Note that this bug did not affect fast path users, since the
fast path does not return poses for camera packets.

rc_getPose now returns an rc_PoseTime with the time corresponding to
the pose (in the case of a pose derived from a camera packet, this is
the center of the exposure).