transformation: implement an EPnP based version of estimate_transform
This version is faster than the version from OpenCV, at least for
small problems, and though it uses the same underlying algorithm and
though it is implemented with floats instead of doubles, it's usually
(98.6% of the time in my tests) more accurate.

The main advantage, numerically, seems to be computing the SVD of M
instead not M'M in order to get the right singular vectors which
avoids the square and sqrt.