covariance: reduce MYRIAD statesize to 80 for now
In practice, the version we have been running on TM2 effectively was
also limited to 80 due to a quirk where we had a minimum group size
of 16 and a maximum group size of 20, so when we were around 80+ we
would be unable to add a group. Due to the smaller maximum size, we
would have more groups, and the fixed group size overhead would often
push us into this corner case.

This makes performance match what we already have on movidius. We
should look to increase this again as soon as possible, after we have
added further optimization on the matrix operations.