rc_tracker: only decimate the slow path
This allows for fewer updates on the slow path, while maintaining low
latency on the fast path.  We can no longer modify the measuremt
variance for the sensor directly, so effect of the effect is open
coded on the slow path.

Additionally, moving decimation to filter_*_measurement and copying
the data before modifying it ensures that we capture the undecimated
data.