filter: Set the intrinsics properly for the second camera in stereo
With monocular cameras, our filter_image_measurement sets the
intrinsics of the camera to match the width and height of the image
which we pass in as opposed to the resolution specified in the
calibration. In stereo this gets updated for camera 0 via
filter_image_measurement but not set for camera1.