filter: Disable bias estimation when we lose all features for 500ms
When we enter orientation only mode and DynamicInitialization run
state, we will naturally not update the biases. But when we recover
feature tracking, we will assume we are not moving and can incorporate
a huge motion into our biases, leading to flyaway.

This workaround suggests that by the time we get into orientation only
mode we should assume we have well estimated the biases and should
trust them when we recover feature tracking.