mavtocapture: add decimate_by: 5 to accelerometer
We don't behave well with undecimated accelerometer with the
measurement variance specified by the yaml (spikes cause us to
immediately diverge). An alternative to this would be to increase the
accelerometer measurement standard deviation *5, but this way should be
easier to remove if we can resolve how to deal with spikes
appropriately.