mapper: the the graph to find closest node to the stage(s), not its old world origin
This fixes lots the threading issues, simplifies and clarifies the
rc_*Stage API, supports multiple stages, removes the default stage and
uses the graph to find the shortest path to the stage to minimize the
error.

The rc_StageCallback is current being called on each relocalization.

rc_getStage will currently only return stages that are on the current
map, not loaded ones and the graph is currently not being connected
after relocalization.